yml code block rendering issue in dendron.topic.hierarchy.md
Hi,
the yml code blocks starting on line 67 are not rendered properly. I believe this is due to the starting `-` in the yml code.  I can't verify this, as the issue doesn't occur in the preview on Github. I just removed the bullet points. Thanks for your work on this project. I'm just trying it out, it looks really cool.