Help beginners with Note References
Added 
```
- BE CAREFUL - some methods of selecting the line will put an extra space after the work "more!". We don't want that space. Make sure that what's highlighted is just  the text from "##" to "more!"
```

Selecting the header for a note reference has to be done carefully so it doesn't include an extra char at the end of the line. Triple clicking includes the extra character and CMD-SPACE-R will just give ![[the_note]], not ![[the_not#with-the-header,1:#*]].
the VIM emulation (dd + P) does this too.
CTRL-CMD-SPACE and right-arrow x 2 doesn't (just selects the text)
I have a short screen cast if that would help.