fix: remove import to missing schema
This caused a "schema malformed" error. In a previous commit `pro.schema.yml` has been renamed to `pkg.schema.yml`. 
Meaning also this  error happpens when an schema import does not exist. The error message did not include the schema file that was responsible for that which made it non-trivial to find out where to look for. We might want to allow this kind of error like @kevins8 mentioned and only message the user about it.