Update dendron.topic.publishing.quickstart.publishing-your-site.md
Add the instruction to use command `Dendron: Worksapce: Sync` next to `git add commit push`  chain in the last step. I'm not sure whether it simplify or give too much information for new user. But actually, as a non-tech user, I prefer to learn the existence of the command `Dendron: Worksapce: Sync` which is much simpler and intuitive to use