Update dendron.ref.cli.md
Updated to include all new flags for the commands based on output on `--help` call, and replaced legacy publishing command with new nextjs one. 

Link to publish page not actually tested as I made this change in the github GUI.

Note that I haven't tested the example for the `convert` action on the vault commands section, and did not know what to put for the info and removeCache options for `workspace` so have left these headers commented out for now until they can be populated.